move the access-token routes into their own namespace to simplify the api

+++ routes/web.php
@@ -42,12 +42,11 @@ $router->group(['middleware' => ['cors']], function() use ($router) {
             $router->post('/',                          'UserController@create');
             $router->put('/{id:[0-9]+}',                'UserController@update');
             $router->delete('/{id:[0-9]+}',             'UserController@deleteById');
+        });
 
-            $router->post('/{id:[0-9]+}/token',         'AccessTokenController@add');
-            $router->delete(
-                '/{userId:[0-9]+}/token/{tokenId:[0-9]+}',
-                'AccessTokenController@remove'
-            );
+        $router->group(['prefix' => 'access-token'], function() use ($router){
+            $router->post('/{userId:[0-9]+}',                       'AccessTokenController@add');
+            $router->delete('/{userId:[0-9]+}/{tokenId:[0-9]+}',    'AccessTokenController@remove');
         });
 
         $router->group(['prefix' => 'package-group'], function() use ($router) {
